Added option tests for TooGenericExceptionCaught
TooGenericExceptionCaughtSpec and TooGenericExceptionThrownSpec were
splitted to make it more readable.
