+++ detekt-rules/src/test/resources/cases/TooGenericExceptionsOptions.kt
@@ -1,0 +1,16 @@
+@file:Suppress("unused")
+
+package cases
+
+class TooGenericExceptionsOptions {
+
+	fun f() {
+		try {
+			throw Throwable()
+		} catch (myIgnore: MyTooGenericException) {
+			throw Error()
+		}
+	}
+}
+
+class MyTooGenericException : RuntimeException()
